Add option to delete worlds (closes #46)
Added `WorldDeleteDialogComponent.vue` - a button which opens a dialog that prompts the user to type in the name of the selected server to prevent accidental deletions. If the correct name is entered the user is able to press the `DELETE` button, which sends a message to the backend.

The backend then removes the world (after shutting the server down) based on the name defined in `server.properties` and brings the server back into it's previous state (i.e. starts the server, if it was started before the world was deleted).