Host without proxy (closes #113)
The proxy is removed and the backend handles all requests in a production environment. In development mode the frontend proxies the websocket requests to the backend.

In production mode the backend now listens on port 8081, in development mode the frontend listens on 8081 and the backend on port 3001.

Reduces the image size from about 490MB to around 340MB.