ci: cache `node_modules` if exact match
Previously only the npm cache was cached. Even though the cache is restored, `npm install` runs for quite a long time on every workflow. Caching `node_modules` improves on this: If an exact cache hit occurs, running `npm install` is not necessary.

Precautions have to be taken if the workflow is supposed to be run across multiple node versions, as dependencies might not work in other versions that the one they were installed in.