docs: state more precisely how to set the server path (#146)
State more precisely that the server path (relative or absolute) can be set using the environment variable `SERVER_PATH` and that the default value is the directory `servers/` relative to the blockcluster root directory. Using `settings.properties` is deprecated and the option will be removed in `v1.0.0` and is therefore not mentioned in `README.md`.