ci: don't `[skip ci]` on release commits
Skipping CI on these commits means that fast forwarding `master` to the latest release commit on `beta` and pushing, creates no new release. Therefore it is important that CI runs on these commits as well.

Also it seems to make more sense, if the `master` and `beta` branch tags as well as the commit sha tags on dockerhub are always up to date with their released versions (even though it's just the `**/package*.json`s that changed).