build: upgrade to java version 17
Upgrade the primary java version to java 17. Java 17 has reached general availability and can therefore be used in a production build.

As some flavours of minecraft servers (e.g. spigot) strictly require java 16 (both as a minimum and maximum) for minecraft version 1.17, it might not be possible to run those servers. If this is necessary additional JREs/ JDKs can be mounted in `/opt/java-xx`. For more details see #150.