feat: allow the selection of a specific java runtime
Allows the selection of a specific java runtime in the frontend. The frontend receives a list of all available java runtimes from the backend as well as the currently selected or default runtime for the current server. The user is able to select another runtime. This information is stored in the `blockcluster.cfg` file in the respective server directory and persisted over restarts of blockcluster. The selected runtime is then used to start the respective server's jar.