feat: allow multiple java versions to be installed (#209)
Allows multiple java runtimes to be installed and the selection of a specific runtime in the frontend.

Includes the LTS versions 11 and 17 of java, which are capable to run the vast majority of minecraft servers. Spigot 1.17 servers strictly require java 16 and are therefore not going to run (vanilla servers of all versions should be able to run on java 17 though). Spigot 1.18 should work on java 17 as well. For more details and instructions to use custom java runtimes see #150.

The frontend receives a list of all discovered java runtimes from the backend as well as the currently selected or default runtime for the current server. The user is able to select another runtime. This information is stored in the `blockcluster.cfg` file in the respective server directory and persisted over restarts of blockcluster. The selected runtime is then used to start the respective server's jar when requested.

Closes #150