refactor: remove workaround for GitHub icon size
vuetifyjs/vuetify#13412 fixed a bug that `v-icon`s didn't adjust to their `v-chip`'s size. Therefore, the icon was rendered as small. A new class was added as a workaround to render the icon in its intended size.

vuetifyjs/vuetify#13813 reverts that bug fix, as "it breaks more common use cases" than it fixed. Therefore, the workaround can be removed.