build: upgrade `alpine` base image to version 3.15
Upgrades the `alpine` base image to version 3.15. This image installs version 16 of node (and version 8 of npm) by default, when running `apk add --update nodejs npm`.

It was assumed that these versions of node and npm were already in use, but that was not the case. Only the builder used version 16 previously.

This change removes a warning during the build process of the docker image.