fix: send status updates while being stopped
The backend now keeps sending updates on the server status as well as console messages until all servers are stopped.

Makes the `Server.stop()` method truly asynchronous and therefore allows multiple event listeners listening on the `exit` event to be removed. The promise is only resolved when the server is entirely stopped. Calling the `stop()` method now also sends the updated status `Stopping` directly to the frontend. Once the server is stopped a second status update is sent to the frontend.

Fixes #232