fix: disable "SAVE" button if flags are unmodified
Disables the button that saves the flags as long as the flags are unmodified. This has multiple benefits: Firstly, the user receives a visual indication that the value is not yet saved or applied. Also, it becomes clear that saving has no effect if the previous flags are matched. Finally, it provides a visual indication that updates from the backend will not be applied due to the changes in the flags field.