ci: run build workflow on `updates` branch
Runs the build workflow on every push to the `updates` branch and for each pull requests with the `updates` branch as its base.

As the repository is public there is no billing for using GitHub Actions anymore. Therefore, it is reasonable to run continuous integration on these events, and it's going to simplify the review of pull requests for new dependency version upgrades by dependabot.