feat(gui): show player count and faces in the frontend
Show player count and faces in the frontend. The player stats are received for online servers in the backend. Afterwards for each online player in the sample the skin is retrieved using https://github.com/jojomatik/pixelheads from the Mojang API. This data is then sent to the frontend with the current server status.

The player count and faces (and other information on the server status) are updated, if a join or leave message is retrieved when a player joins or leaves a server. Therefore, it is not necessary to continuously update the status (see also #59, where an event based system was preferred over continuous status updates). A timeout is specified, before player information is updated as the sample and count are not updated immediately on the server.