fix(backend): reliably update player sample
Update player sample reliably after a player leaves the server. 5 seconds seems to be an interval after which the server reliably performs a recalculation of the player sample.

Before this change it occasionally happened that player faces weren't unrendered in the frontend even after they left the server. This behaviour was never observed after a player joined, which leads to assumption that 5 seconds is a reliable timeout to use.