perf(backend): don't block status update with I/O-operations
Don't block status update with I/O-operations by calling `Server.update()` when all servers are queried in one batch. Instead, only call `Server.updateStatus()` which is sufficient for the relevant use case in the server overview. `Server.update()` is still called when a server is queried separately from a detailed `ServerView`.

The previous approach lead to long response times for instances with many servers, as the `blockcluster.cfg` was read for each of them and as file system read operations are seemingly handled subsequently instead of in parallel.

Fixes #280