build(deps): recreate `package-lock.json`
Recreate `package-lock.json` by deleting it and `node_modules/` and running `npm install` to remove optional peer dependencies on `webpack` 4.

Some relevant changes include: Remove `cache-loader` (optional dependency) as it depends on `webpack` 4 and therefore blocks the "global" upgrade to `webpack` 5, remove `terser-webpack-plugin` as it circularly depends on `webpack` 4 (which depends on `terser-webpack-plugin`) and therefore blocks the "global" upgrade to `webpack` 5.