feat: add option to pause servers on idle
Add option to automatically pause servers on idle (i.e. with no players online). Add a timeout that starts if no players are online and stops the server after a configurable amount of time. If a player joins during that time the timeout is cleared and the server won't stop. While a server is paused a mock-server based on `minecraft-protocol` is started and handles connections by clients. It shows the status in the minecraft client server overview and if a player connects the server will start and the player is prompted to reconnect in a few seconds.

This feature can be configured in the frontend or by manually configuring it in `blockcluster.cfg` for each server. By default, this feature is disabled and the timeout is set to 300s.

Changes to the server status and idle timeout are being logged to the server itself as well as the blockcluster frontend.

Closes #279