fix: don't crash backend if selected java runtime is not available
Don't crash backend if selected java runtime is not available by checking if it is in the list of available runtimes and throwing a custom error, if this is not the case. Handle the error by logging it to the console in the relevant locations and sending the error message to the frontend.

Disable `StateChangeButtonComponent`s in the frontend, if the java runtime is not included in the list of available java runtimes.

Fixes #386