feat: show resource usage of backend in `HomeView`
Show the resource usage of the blockcluster backend in the total resource usage chart in `HomeView`.

This includes the resource usage of paused servers listening for connections. As the resource usage of the backend cannot be allocated to individual servers it is only shown in the `HomeView` and not individually for each server.

Closes #360