build(deps): downgrade vue from 2.7.4 to 2.6.14
Downgrades [vue](https://github.com/vuejs/vue) from 2.7.4 to 2.6.14.
- [Release notes](https://github.com/vuejs/vue/releases)
- [Changelog](https://github.com/vuejs/vue/blob/main/CHANGELOG.md)
- [Commits](https://github.com/vuejs/vue/commits)

This downgrade was done as the vue 3 backport caused some issues in the docker development environment (errors in the browser console) and in other projects. This version of vue will probably be skipped entirely. As soon as we migrate to nuxt 3 and vue 3 this should be irrelevant anyway.

Downgrades [vue-template-compiler](https://github.com/vuejs/vue) from 2.7.4 to 2.6.14.
- [Release notes](https://github.com/vuejs/vue/releases)
- [Changelog](https://github.com/vuejs/vue/blob/main/CHANGELOG.md)
- [Commits](https://github.com/vuejs/vue/commits)

Downgrades [vuetify-loader](https://github.com/vuetifyjs/vuetify-loader) from 1.9.1 to 1.8.0.
- [Release notes](https://github.com/vuetifyjs/vuetify-loader/releases)
- [Commits](https://github.com/vuetifyjs/vuetify-loader/commits)

---
updated-dependencies:
- dependency-name: vue
  dependency-type: direct:production
  update-type: version-update:semver-minor
- dependency-name: vue-template-compiler
  dependency-type: direct:development
  update-type: version-update:semver-minor
- dependency-name: vuetify-loader
  dependency-type: direct:development
  update-type: version-update:semver-minor
...