build(backend): use local typescript installation
Use a local installation of typescript to compile the backend instead of the global installation in the `Dockerfile`. This has the benefit, that the version defined in `package.json` is used instead of the newest available typescript version on npm. Update documentation to reference the new build script `npm run build` instead of `tsc`.